+++ cloudant-sync-datastore-core/src/main/java/com/cloudant/sync/internal/replication/PullStrategy.java
@@ -236,9 +236,10 @@ public class PullStrategy implements ReplicationStrategy {
         }
 
         this.state.documentCounter = 0;
+
         while (!this.state.cancel) {
             this.state.batchCounter++;
-
+            final Object lastKnownCheckpoint = this.targetDb.getCheckpoint(this.getReplicationId());
             String msg = String.format(
                     "Batch %s started (completed %s changes so far)",
                     this.state.batchCounter,
@@ -247,7 +248,7 @@ public class PullStrategy implements ReplicationStrategy {
             logger.info(msg);
             long batchStartTime = System.currentTimeMillis();
 
-            ChangesResultWrapper changeFeeds = this.nextBatch();
+            ChangesResultWrapper changeFeeds = this.nextBatch(lastKnownCheckpoint);
             int batchChangesProcessed = 0;
 
             // So we can check whether all changes were processed during
@@ -262,6 +263,16 @@ public class PullStrategy implements ReplicationStrategy {
             if (changeFeeds.size() > 0) {
                 batchChangesProcessed = processOneChangesBatch(changeFeeds);
                 state.documentCounter += batchChangesProcessed;
+            }
+
+            if (!this.state.cancel && (lastKnownCheckpoint == null || !lastKnownCheckpoint.equals(changeFeeds.getLastSeq()))) {
+                try {
+                    this.targetDb.putCheckpoint(this.getReplicationId(), changeFeeds.getLastSeq());
+                } catch (DocumentStoreException e) {
+                    logger.log(Level.WARNING, "Failed to put checkpoint doc, next replication " +
+                            "will " +
+                            "start from previous checkpoint", e);
+                }
             }
 
             long batchEndTime = System.currentTimeMillis();
@@ -435,15 +446,6 @@ public class PullStrategy implements ReplicationStrategy {
             }
         }
 
-        if (!this.state.cancel) {
-            try {
-                this.targetDb.putCheckpoint(this.getReplicationId(), changeFeeds.getLastSeq());
-            } catch (DocumentStoreException e) {
-                logger.log(Level.WARNING, "Failed to put checkpoint doc, next replication will " +
-                        "start from previous checkpoint", e);
-            }
-        }
-
         return changesProcessed;
     }
 
@@ -470,8 +472,7 @@ public class PullStrategy implements ReplicationStrategy {
         }
     }
 
-    private ChangesResultWrapper nextBatch() throws DocumentStoreException {
-        final Object lastCheckpoint = this.targetDb.getCheckpoint(this.getReplicationId());
+    private ChangesResultWrapper nextBatch(final Object lastCheckpoint) {
         logger.fine("last checkpoint " + lastCheckpoint);
 
         ChangesResult changeFeeds = null;