+++ cloudant-sync-datastore-core/src/test/java/com/cloudant/sync/internal/replication/PullStrategyMockTest.java
@@ -16,6 +16,8 @@
 
 package com.cloudant.sync.internal.replication;
 
+import static org.junit.Assume.assumeNoException;
+import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.anyInt;
 import static org.mockito.Mockito.anyString;
@@ -26,12 +28,12 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.cloudant.common.RequireRunningCouchDB;
+import com.cloudant.sync.event.Subscribe;
+import com.cloudant.sync.internal.documentstore.DocumentRevsList;
 import com.cloudant.sync.internal.mazha.ChangesResult;
 import com.cloudant.sync.internal.mazha.DocumentRevs;
 import com.cloudant.sync.internal.mazha.OkOpenRevision;
 import com.cloudant.sync.internal.mazha.OpenRevision;
-import com.cloudant.sync.internal.documentstore.DocumentRevsList;
-import com.cloudant.sync.event.Subscribe;
 import com.cloudant.sync.internal.util.JSONUtils;
 import com.cloudant.sync.replication.PullFilter;
 import com.cloudant.sync.util.TestUtils;
@@ -230,6 +232,107 @@ public class PullStrategyMockTest extends ReplicationTestBase {
         verify(mockListener).complete(any(ReplicationStrategyCompleted.class));
         verify(mockListener,never()).error(any(ReplicationStrategyErrored.class));
     }
+
+    @Test
+    public void testSetCheckpointWhenEmpty() throws Exception {
+        CouchDB mockRemoteDb = mock(CouchDB.class);
+        when(mockRemoteDb.changes((PullFilter) null, null, 1000)).then(new Answer<Object>() {
+            @Override
+            public Object answer(InvocationOnMock invocation) throws Throwable {
+                FileReader fr = new FileReader(TestUtils.loadFixture
+                        ("fixture/empty_changes.json"));
+                return JSONUtils.fromJson(fr, ChangesResult.class);
+            }
+        });
+        when(mockRemoteDb.exists()).thenReturn(true);
+
+        StrategyListener mockListener = mock(StrategyListener.class);
+        PullStrategy pullStrategy = super.getPullStrategy();
+        pullStrategy.sourceDb = mockRemoteDb;
+        pullStrategy.getEventBus().register(mockListener);
+        pullStrategy.run();
+
+        //should have 0 document
+        Assert.assertEquals(this.datastore.getDocumentCount(), 0);
+        //Checkpoint should be created in targetDb
+        String checkpoint =
+                (String) pullStrategy.targetDb.getCheckpoint(pullStrategy.getReplicationId());
+        Assert.assertEquals(checkpoint, "10-d9e5b0147af143e5b6d1979378ad957b");
+        //make sure the correct events were fired
+        verify(mockListener).complete(any(ReplicationStrategyCompleted.class));
+        verify(mockListener, never()).error(any(ReplicationStrategyErrored.class));
+    }
+
+    @Test
+    public void testDoNotSetCheckpointWhenNotModified() throws Exception {
+        try {
+            CouchDB mockRemoteDb = mock(CouchDB.class);
+            when(mockRemoteDb.changes((PullFilter) null, "10-d9e5b0147af143e5b6d1979378ad957b", 1000)).then(new Answer<Object>() {
+                @Override
+                public Object answer(InvocationOnMock invocation) throws Throwable {
+                    FileReader fr = new FileReader(TestUtils.loadFixture
+                            ("fixture/empty_changes.json"));
+                    return JSONUtils.fromJson(fr, ChangesResult.class);
+                }
+            });
+            when(mockRemoteDb.exists()).thenReturn(true);
+
+            DatastoreWrapper mockLocalDb = mock(DatastoreWrapper.class);
+            when(mockLocalDb.getCheckpoint(any(String.class))).thenReturn("10" +
+                    "-d9e5b0147af143e5b6d1979378ad957b");
+
+            StrategyListener mockListener = mock(StrategyListener.class);
+            PullStrategy pullStrategy = super.getPullStrategy();
+            pullStrategy.sourceDb = mockRemoteDb;
+            pullStrategy.targetDb = mockLocalDb;
+            pullStrategy.getEventBus().register(mockListener);
+            pullStrategy.run();
+
+            //make sure the correct events were fired
+            verify(mockListener).complete(any(ReplicationStrategyCompleted.class));
+            verify(mockListener, never()).error(any(ReplicationStrategyErrored.class));
+            verify(mockLocalDb, never()).putCheckpoint(anyString(), any());
+        } catch(UnsupportedOperationException uoe) {
+            assumeNoException("Cannot proxy DatastoreWrapper on Dalvik", uoe);
+        }
+    }
+
+    @Test
+    public void testSetCheckpointWhenModified() throws Exception {
+        try {
+            CouchDB mockRemoteDb = mock(CouchDB.class);
+            when(mockRemoteDb.changes((PullFilter) null, "9-d9e5b0147af143e5b6d1979378ad957b",
+                    1000)).then(new Answer<Object>() {
+                @Override
+                public Object answer(InvocationOnMock invocation) throws Throwable {
+                    FileReader fr = new FileReader(TestUtils.loadFixture
+                            ("fixture/empty_changes.json"));
+                    return JSONUtils.fromJson(fr, ChangesResult.class);
+                }
+            });
+            when(mockRemoteDb.exists()).thenReturn(true);
+
+            DatastoreWrapper mockLocalDb = mock(DatastoreWrapper.class);
+            when(mockLocalDb.getCheckpoint(any(String.class))).thenReturn("9" +
+                    "-d9e5b0147af143e5b6d1979378ad957b");
+
+            StrategyListener mockListener = mock(StrategyListener.class);
+            PullStrategy pullStrategy = super.getPullStrategy();
+            pullStrategy.sourceDb = mockRemoteDb;
+            pullStrategy.targetDb = mockLocalDb;
+            pullStrategy.getEventBus().register(mockListener);
+            pullStrategy.run();
+
+            //make sure the correct events were fired
+            verify(mockListener).complete(any(ReplicationStrategyCompleted.class));
+            verify(mockListener, never()).error(any(ReplicationStrategyErrored.class));
+            verify(mockLocalDb).putCheckpoint(anyString(), eq("10" +
+                    "-d9e5b0147af143e5b6d1979378ad957b"));
+        } catch (UnsupportedOperationException uoe) {
+            assumeNoException("Cannot proxy DatastoreWrapper on Dalvik", uoe);
+        }
+    }
+
 
     public class StrategyListener {